Diabetic-Retinopathy/requirements.txt
albumentations==1.2.1
dill==0.3.4
imbalanced_learn==0.9.1
imblearn==0.0
imutils==0.5.4
joblib==1.1.0
matplotlib==3.5.2
natsort==8.1.0
numpy==1.22.3
opencv_python_headless==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
scikit-learn
scipy==1.8.1
seaborn==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
wandb==0.12.21